# SourceCode/swagger.yml
...
openapi: 3.0.0
info:
  title: "Trail microservice API"
  description: "An API for trails"
  version: "1.0.0"
servers:
  - url: "/api"
components:
  schemas:
...
    Feature:
      type: "object"
      properties:
        FeatureId:
          type: "integer"
        Feature:
          type: "string"
    TrailFeature:
      type: "object"
      properties:
        FeatureId:
          type: "integer"
        TrailId:
          type: "integer"
...
  parameters:
...
    FeatureId:
      name: "FeatureId"
      description: "Id of the feature"
      in: path
      required: True
      schema:
        type: "integer"
...
paths:
...
  /features:
    get:
      operationId: "features.getFeatures"
      tags:
        - Features
      summary: "Gets all features"
      responses:
        "200":
          description: "OK"
        "500":
          description: "Server error"

    post:
      operationId: "features.addFeature"
      tags:
        - Features
      summary: "Add new feature"
      requestBody:
        description: "Feature to add"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "pointJSON"
              $ref: "#/components/schemas/Feature"
      responses:
        "201":
          description: "Successfully added feature"
        "400":
          description: "Bad request"
        "500":
          description: "Server error"

    put:
      operationId: "features.editFeatureById"
      tags:
        - Features
      summary: "Update feature"
      requestBody:
        description: "Feature to update"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "featureJSON"
              $ref: "#/components/schemas/Feature"
      responses:
        "201":
          description: "Successfully updated feature"
        "400":
          description: "Bad request"
        "404":
          description: "Feature not found"
        "500":
          description: "Server error"
  /features/{featureId}:

    get:
      operationId: "features.getFeaturePointById"
      tags:
        - Features
      summary: "Gets feature with given Id"
      parameters:
       - $ref: "#/components/parameters/FeatureId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "Feature not found"
        "500":
          description: "Server error"


    delete:
      operationId: "features.deleteFeatureById"
      tags:
        - Features
      summary: "Deletes feature with given id"
      parameters:
       - $ref: "#/components/parameters/FeatureId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "Feature not found"
        "500":
          description: "Server error"

# Features
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
# Location Points
...
  /trailfeatureslink:
    post:
      operationId: "linkhelper.newTrailFeature"
      tags:
        - Link entity
      summary: "Add new trail-feature link"
      requestBody:
        description: "Feature and trail to link"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "TrailFeatureJSON"
              $ref: "#/components/schemas/TrailFeature"
      responses:
        "201":
          description: "Successfully added trail feature"
        "400":
          description: "Bad request"
        "500":
          description: "Server error"

    delete:
      operationId: "linkhelper.deleteTrailFeature"
      tags:
        - Link entity
      summary: "Delete trail-feature link"
      requestBody:
        description: "Feature and trail to delete link of"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "TrailFeatureJSON"
              $ref: "#/components/schemas/TrailFeature"
      responses:
        "201":
          description: "Successfully deleted trail feature"
        "400":
          description: "Bad request"
        "404":
          description: "No TrailFeature found"
        "500":
          description: "Server error"
...